+++ apps/api-service/src/shared/constants/db-connection.map.ts
@@ -7,6 +7,6 @@ export enum ProductsDbConnectionMap {
   WealthAnalyzer = "wealthanalyzer",
   DebtTrack = "debttrack",
   WealthGoal = "wealthgoal",
-  Intelligence = "intelligence",
+  OneAgent = "oneagent",
   TaxAdvisor = "taxadvisor",
 }