+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 ## v1.7
 
 - Dependency updates & bug fixes
+- Minor UI refactoring
+- Rebranded Intelligence to OneAgent
+- Standardized date format
+
+## v1.7
+
+- Dependency updates & bug fixes
 - Code optimization
 
 ## v1.6